+++ examples/data-sources/iosxe_static_route_vrf/data-source.tf
@@ -1,3 +1,0 @@
-data "iosxe_static_route_vrf" "example" {
-  vrf = "VRF1"
-}