+++ examples/data-sources/iosxe_static_routes_vrf/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxe_static_routes_vrf" "example" {
+  vrf = "VRF1"
+}