+++ examples/resources/iosxe_static_route_vrf/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxe_static_route_vrf.example "<vrf>"