+++ examples/resources/iosxe_static_routes_vrf/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxe_static_routes_vrf.example "<vrf>"